+++ src/synchronisation/base/DummySync.cc
@@ -36,7 +36,7 @@ void DummySync::initialize(int stage)
         Scheduled::initialize();
 
         SchedulerActionTimeEvent *event = new SchedulerActionTimeEvent("Sync Task Event", ACTION_TIME_EVENT);
-        event->setAction_time(par("action_time").longValue());
+        event->setAction_time((uint32_t)par("action_time").longValue());
         event->setDestinationGate(gate("schedulerIn"));
         period->registerEvent(event);
     }
@@ -55,7 +55,7 @@ void DummySync::handleMessage(cMessage *msg)
         period->registerEvent(event);
 
         if(period->getCycles()>0){
-            uint32_t cycleTicks = period->par("cycle_ticks").longValue();
+            uint32_t cycleTicks = (uint32_t)period->par("cycle_ticks").longValue();
             simtime_t tick = oscillator->par("tick").doubleValue();
 
             int64_t modticks = ((int64_t)(simTime()/tick)-par("action_time").longValue())%cycleTicks;
@@ -63,7 +63,7 @@ void DummySync::handleMessage(cMessage *msg)
                 modticks=modticks-cycleTicks;
             modticks+=uniform(-par("precission").doubleValue()/2, par("precission").doubleValue()/2)/tick;
 
-            timer->clockCorrection(modticks);
+            timer->clockCorrection((int32_t)modticks);
         }
     }
 }
